+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/hooks/submissions/use-submissions-details.tsx
@@ -1,11 +1,15 @@
 import React, { createContext, useCallback, useContext, useEffect, useMemo, useState } from 'react';
 import { ITestRunType, ISubmissionType, ISubmissionDetailsType, ISubmissionResultType } from './types';
+
 import { getSubmissionDetailsUrl, getSubmissionResultsByProblem } from '../../utils/urls';
+
+import { IHaveChildrenProps } from '../../components/common/Props';
+
 import { useLoading } from '../use-loading';
 import { useHttp } from '../use-http';
-import { useContests } from '../contests/use-contests';
+import { useContests } from '../use-contests';
+
 import DEFAULT_PROBLEM_RESULTS_TAKE_CONTESTS_PAGE from '../../common/constants';
-import { IHaveChildrenProps } from '../../components/common/Props';
 
 interface ISubmissionsDetailsContext {
     setCurrentSubmissionId: (submissionId: number) => void;
